pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zodb-s3blobs"
version = "0.1.0.dev0"
description = "Store ZODB blobs in S3-compatible object storage"
license = "ZPL-2.1"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Zope Public License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
]
dependencies = [
    "ZODB>=5.8",
    "boto3>=1.26",
    "zope.interface",
    "zc.lockfile",
    "ZConfig",
    "transaction",
]

[project.urls]
Homepage = "https://github.com/bluedynamics/zodb-s3blobs"
Repository = "https://github.com/bluedynamics/zodb-s3blobs"
Issues = "https://github.com/bluedynamics/zodb-s3blobs/issues"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "moto[s3]>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/zodb_s3blobs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=zodb_s3blobs --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
source = ["zodb_s3blobs"]

[tool.coverage.report]
show_missing = true
fail_under = 90
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "A",    # flake8-builtins (shadowing builtins)
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    # PTH omitted: ZODB APIs use string paths throughout
    "RUF",  # ruff-specific
    "SIM",  # flake8-simplify
    "T20",  # flake8-print
    "UP",   # pyupgrade
    "W",    # pycodestyle warnings
]
ignore = [
    "E501",   # line too long
    "RUF001", # Docstring contains ambiguous `’`
    "RUF002", # Docstring contains ambiguous `’`
]

[tool.ruff.lint.isort]
# plone style
force-single-line = true
from-first = true
lines-after-imports = 2
lines-between-types = 1
no-sections = true
order-by-type = false

[tool.ruff.lint.pycodestyle]
max-line-length = 120  # E501 triggers above this, allows flexibility
max-doc-length = 120  # W505 triggers above this

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T201"]

[tool.ty.rules]
unsupported-base = "ignore"

